Refactor various utility classes for improved type handling and error management

+++ classes/Utils/CSS.php
@@ -35,13 +35,13 @@ class PUM_Utils_CSS {
 		}
 
 		if ( strlen( $hex ) === 3 ) {
-			$r = hexdec( substr( $hex, 0, 1 ) . substr( $hex, 0, 1 ) );
-			$g = hexdec( substr( $hex, 1, 1 ) . substr( $hex, 1, 1 ) );
-			$b = hexdec( substr( $hex, 2, 1 ) . substr( $hex, 2, 1 ) );
+			$r = (int) hexdec( substr( $hex, 0, 1 ) . substr( $hex, 0, 1 ) );
+			$g = (int) hexdec( substr( $hex, 1, 1 ) . substr( $hex, 1, 1 ) );
+			$b = (int) hexdec( substr( $hex, 2, 1 ) . substr( $hex, 2, 1 ) );
 		} else {
-			$r = hexdec( substr( $hex, 0, 2 ) );
-			$g = hexdec( substr( $hex, 2, 2 ) );
-			$b = hexdec( substr( $hex, 4, 2 ) );
+			$r = (int) hexdec( substr( $hex, 0, 2 ) );
+			$g = (int) hexdec( substr( $hex, 2, 2 ) );
+			$b = (int) hexdec( substr( $hex, 4, 2 ) );
 		}
 
 		$rgb = [ $r, $g, $b ];
